+++ js/common/maps.js
@@ -1,0 +1,10 @@
+import { HomeType } from './enums.js';
+
+const homeTypeToReadable = {
+  [HomeType.FLAT]: 'Квартира',
+  [HomeType.BUNGALOW]: 'Бунгало',
+  [HomeType.HOUSE]: 'Дом',
+  [HomeType.PALACE]: 'Дворец',
+};
+
+export { homeTypeToReadable };
